requirements.txt
amqp==2.2.1
billiard==3.5.0.3
boto3==1.4.7
botocore==1.7.0
cachetools==2.0.1
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
dill==0.2.7.1
Django==1.11.4
docutils==0.14
enum34==1.1.6
future==0.16.0
futures==3.1.1
google-auth==1.0.2
google-cloud-core==0.27.1
google-cloud-pubsub==0.28.2
google-gax==0.15.14
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.1
grpcio==1.6.0
httplib2==0.10.3
idna==2.6
jmespath==0.9.3
kombu==4.1.0
oauth2client==3.0.0
ply==3.8
protobuf==3.4.0
psutil==5.3.0
pyasn1==0.3.3
pyasn1-modules==0.1.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
urllib3==1.22
vine==1.1.4
